requirements.txt
gradio
python-dotenv
langchain-openai
langchain-ollama
langchain-anthropic
rich
